feat: Phase 1 - CSR + CORS 設定による API 連携完了
- フロントに CSR 版の投稿一覧ページを実装（useEffect + fetch）
- CORSエラーを意図的に発生させて観察
- Spring Boot に CORS 設定を追加して解決
- Entity の id 型を Long → Integer に修正（SERIAL との整合性）
- learning-log に React/Spring Boot の学習内容を追記

+++ api/src/main/java/com/example/api/entity/Post.java
@@ -22,10 +22,14 @@ public class Post {
     /**
      * 主キー
      * GenerationType.IDENTITY = DB の自動採番（SERIAL）を使用
+     *
+     * 【学習ポイント】
+     * PostgreSQL の SERIAL は INTEGER（32ビット）。
+     * Java の Long は BIGINT（64ビット）に対応するため、型を合わせる必要がある。
      */
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     /**
      * タイトル（必須）
@@ -59,11 +63,11 @@ public class Post {
     // JPA はリフレクションでフィールドにアクセスするため、
     // Getter/Setter がなくても動作するが、慣例として用意する。
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
